requirements.txt
matplotlib==3.7.1
numpy==1.24.2
opencv_python==4.1.2.30
opencv_python_headless==4.7.0.72
pandas==1.5.3
Pillow==6.2.2
Pillow==9.5.0
Pillow==10.3.0
scipy==1.9.1
thop==0.1.1.post2209072238
torch==1.12.0
tqdm==4.65.0
